Aspects

+++ maps/src/main/scala/dsentric/codecs/std/DValueCodecs.scala
@@ -10,7 +10,7 @@ import scala.collection.immutable.SortedSet
 
 trait DValueCodecs {
 
-  implicit val dataCodec:DCodec[Data] =
+  implicit val dataCodec: DCodec[Data] =
     DataCodec
 
   implicit val dValueCodec: DValueCodec[DValue] =
@@ -28,17 +28,17 @@ trait DValueCodecs {
       def typeDefinition: TypeDefinition = TypeDefinition.anyVal
     }
 
-  implicit val stringCodec:DStringCodec[String] with DirectCodec[String] =
+  implicit val stringCodec: DStringCodec[String] with DirectCodec[String] =
     new DStringCodec[String] with DirectCodec[String] {
 
       def fromString(s: String): Option[String] =
         Some(s)
 
-      def typeDefinition:StringDefinition =
+      def typeDefinition: StringDefinition =
         StringDefinition.empty
     }
 
-  implicit val charCodec:DStringCodec[Char] =
+  implicit val charCodec: DStringCodec[Char] =
     new DStringCodec[Char] {
 
       def apply(t: Char): String =
@@ -46,154 +46,156 @@ trait DValueCodecs {
 
       def fromString(s: String): Option[Char] =
         if (s.size == 1)
-           s.headOption
+          s.headOption
         else None
 
       def typeDefinition: StringDefinition =
         StringDefinition(maxLength = Some(1))
     }
 
-  implicit val booleanCodec:DValueCodec[Boolean] =
+  implicit val booleanCodec: DValueCodec[Boolean] =
     new MatchCodec[Boolean] {
       protected def isMatch(a: Raw): Boolean =
         a.isInstanceOf[Boolean]
-      def typeDefinition:TypeDefinition =
+      def typeDefinition: TypeDefinition     =
         BooleanDefinition
     }
 
-  implicit val longCodec:DValueCodec[Long] =
+  implicit val longCodec: DValueCodec[Long] =
     new DValueCodec[Long] with DirectCodec[Long] {
-      def unapply(a:Raw): Option[Long] =
+      def unapply(a: Raw): Option[Long] =
         NumericPartialFunctions.long.lift(a)
 
-      val typeDefinition:TypeDefinition =
+      val typeDefinition: TypeDefinition =
         IntegerDefinition(minimum = Some(Long.MinValue), maximum = Some(Long.MaxValue))
     }
 
-  implicit val doubleCodec:DValueCodec[Double] =
+  implicit val doubleCodec: DValueCodec[Double] =
     new DValueCodec[Double] with DirectCodec[Double] {
-      def unapply(a:Raw): Option[Double] =
+      def unapply(a: Raw): Option[Double] =
         NumericPartialFunctions.double.lift(a)
-      val typeDefinition:TypeDefinition =
+      val typeDefinition: TypeDefinition  =
         NumberDefinition(minimum = Some(Double.MinValue), maximum = Some(Double.MaxValue))
     }
 
-  implicit val intCodec:DValueCodec[Int] =
+  implicit val intCodec: DValueCodec[Int]       =
     new DValueCodec[Int] {
-      def apply(t: Int): Raw =
+      def apply(t: Int): Raw             =
         t.toLong
-      def unapply(a:Raw): Option[Int] =
+      def unapply(a: Raw): Option[Int]   =
         NumericPartialFunctions.int.lift(a)
-      val typeDefinition:TypeDefinition =
+      val typeDefinition: TypeDefinition =
         IntegerDefinition(minimum = Some(Int.MinValue), maximum = Some(Int.MaxValue))
     }
-  implicit val shortCodec:DValueCodec[Short] =
+  implicit val shortCodec: DValueCodec[Short]   =
     new DValueCodec[Short] {
-      def apply(t: Short): Raw =
+      def apply(t: Short): Raw           =
         t.toLong
-      def unapply(a:Raw): Option[Short] =
+      def unapply(a: Raw): Option[Short] =
         NumericPartialFunctions.short.lift(a)
-      val typeDefinition:TypeDefinition =
+      val typeDefinition: TypeDefinition =
         IntegerDefinition(minimum = Some(Short.MinValue), maximum = Some(Short.MaxValue))
     }
-  implicit val byteCodec:DValueCodec[Byte] =
+  implicit val byteCodec: DValueCodec[Byte]     =
     new DValueCodec[Byte] {
-      def apply(t: Byte): Raw =
+      def apply(t: Byte): Raw            =
         t.toLong
-      def unapply(a:Raw): Option[Byte] =
+      def unapply(a: Raw): Option[Byte]  =
         NumericPartialFunctions.byte.lift(a)
-      val typeDefinition:TypeDefinition =
+      val typeDefinition: TypeDefinition =
         IntegerDefinition(minimum = Some(Byte.MinValue), maximum = Some(Byte.MaxValue))
     }
-  implicit val floatCodec:DValueCodec[Float] =
+  implicit val floatCodec: DValueCodec[Float]   =
     new DValueCodec[Float] {
-      def apply(t: Float): Raw =
+      def apply(t: Float): Raw           =
         t.toDouble
-      def unapply(a:Raw): Option[Float] =
+      def unapply(a: Raw): Option[Float] =
         NumericPartialFunctions.float.lift(a)
-      val typeDefinition:TypeDefinition =
-        NumberDefinition(minimum = Some(Float.MinValue), maximum = Some(Float.MaxValue))
+      val typeDefinition: TypeDefinition =
+        NumberDefinition(minimum = Some(Float.MinValue.toDouble), maximum = Some(Float.MaxValue.toDouble))
     }
-  implicit val numberCodec:DValueCodec[Number] =
+  implicit val numberCodec: DValueCodec[Number] =
     new DValueCodec[Number] {
-      def apply(t: Number): Raw =
+      def apply(t: Number): Raw           =
         t.doubleValue()
-      def unapply(a:Raw): Option[Number] =
+      def unapply(a: Raw): Option[Number] =
         NumericPartialFunctions.number.lift(a)
 
-      val typeDefinition:TypeDefinition =
+      val typeDefinition: TypeDefinition =
         NumberDefinition.empty
     }
 
-  implicit def setCodec[T](implicit D:DStringCodec[T]):DValueCodec[Set[T]] =
+  implicit def setCodec[T](implicit D: DStringCodec[T]): DValueCodec[Set[T]] =
     new DValueCodec[Set[T]] {
-      def apply(t: Set[T]): RawObject =
+      def apply(t: Set[T]): RawObject     =
         t.map(D.apply(_) -> 1).toMap
 
       def unapply(a: Raw): Option[Set[T]] =
         a match {
-          case obj:RawObject@unchecked =>
-            obj.foldLeft(Option(Set.empty[T])){
+          case obj: RawObject @unchecked =>
+            obj.foldLeft(Option(Set.empty[T])) {
               case (Some(a), (key, 1)) =>
                 D.unapply(key).map(a + _)
-              case _ =>
+              case _                   =>
                 None
             }
-          case _ =>
+          case _                         =>
             None
         }
 
       def typeDefinition: TypeDefinition = ???
     }
 
-  implicit def nonEmptySetCodec[T](implicit D:DStringCodec[T], O:Ordering[T]):DValueCodec[NonEmptySet[T]] =
+  implicit def nonEmptySetCodec[T](implicit D: DStringCodec[T], O: Ordering[T]): DValueCodec[NonEmptySet[T]] =
     new DValueCodec[NonEmptySet[T]] {
-      def apply(t: NonEmptySet[T]): RawObject =
+      def apply(t: NonEmptySet[T]): RawObject     =
         t.toNonEmptyList.toList.map(D.apply(_) -> 1).toMap
 
       def unapply(a: Raw): Option[NonEmptySet[T]] =
         a match {
-          case obj:RawObject@unchecked if obj.nonEmpty =>
-            obj.foldLeft(Option(SortedSet.empty[T])){
-              case (Some(a), (key, 1)) =>
-                D.unapply(key).map(a + _)
-              case _ =>
-                None
-            }.flatMap(NonEmptySet.fromSet)
-          case _ =>
+          case obj: RawObject @unchecked if obj.nonEmpty =>
+            obj
+              .foldLeft(Option(SortedSet.empty[T])) {
+                case (Some(a), (key, 1)) =>
+                  D.unapply(key).map(a + _)
+                case _                   =>
+                  None
+              }
+              .flatMap(NonEmptySet.fromSet)
+          case _                                         =>
             None
         }
 
       def typeDefinition: TypeDefinition = ???
     }
 
-  implicit val dObjectCodec:DValueCodec[DObject] =
-    new DValueCodec[DObject]{
+  implicit val dObjectCodec: DValueCodec[DObject] =
+    new DValueCodec[DObject] {
 
       def apply(t: DObject): RawValue =
         t.value
 
       def unapply(a: Raw): Option[DObject] =
         a match {
-          case r:RawObject@unchecked =>
+          case r: RawObject @unchecked =>
             Some(new DObjectInst(r))
-          case _ =>
+          case _                       =>
             None
         }
-      def typeDefinition: TypeDefinition =
+      def typeDefinition: TypeDefinition   =
         ObjectDefinition.empty
 
     }
 
-  implicit val dArrayCodec:DValueCodec[DArray] =
+  implicit val dArrayCodec: DValueCodec[DArray] =
     new DValueCodec[DArray] {
       override def apply(t: DArray): RawArray = t.value
 
       def unapply(a: Raw): Option[DArray] =
         a match {
-          case raw: RawArray@unchecked =>
+          case raw: RawArray @unchecked =>
             Some(new DArray(raw))
-          case _ =>
+          case _                        =>
             None
         }
 
@@ -201,51 +203,51 @@ trait DValueCodecs {
         ArrayDefinition(items = Vector(TypeDefinition.anyDefinition))
     }
 
-  implicit val dFilterCodec:DValueCodec[DFilter] =
-    new DValueCodec[DFilter]{
+  implicit val dFilterCodec: DValueCodec[DFilter] =
+    new DValueCodec[DFilter] {
 
       def apply(t: DFilter): RawValue =
         t.value
 
       def unapply(a: Raw): Option[DFilter] =
         a match {
-          case r:RawObject@unchecked =>
+          case r: RawObject @unchecked =>
             Some(new DFilter(r))
-          case _ =>
+          case _                       =>
             None
         }
-      def typeDefinition: TypeDefinition =
+      def typeDefinition: TypeDefinition   =
         ObjectDefinition.empty
     }
 
-  implicit val dProjectionCodec:DValueCodec[DProjection] =
-    new DValueCodec[DProjection]{
+  implicit val dProjectionCodec: DValueCodec[DProjection] =
+    new DValueCodec[DProjection] {
 
       def apply(t: DProjection): RawValue =
         t.value
 
       def unapply(a: Raw): Option[DProjection] =
         a match {
-          case r:RawObject@unchecked =>
+          case r: RawObject @unchecked =>
             Some(new DProjection(r))
-          case _ =>
+          case _                       =>
             None
         }
-      def typeDefinition: TypeDefinition =
+      def typeDefinition: TypeDefinition       =
         ObjectDefinition.empty
     }
 
-  implicit val dNullCodec:DCodec[DNull.type] =
+  implicit val dNullCodec: DCodec[DNull.type] =
     new MatchCodec[DNull.type] {
 
       protected def isMatch(a: Raw): Boolean =
         a.isInstanceOf[DNull.type]
 
-      def typeDefinition:TypeDefinition =
+      def typeDefinition: TypeDefinition =
         NullDefinition
     }
 
-  implicit val pathCodec:DStringCodec[Path] =
+  implicit val pathCodec: DStringCodec[Path] =
     new DStringCodec[Path] {
 
       override def apply(t: Path): String =
@@ -254,7 +256,7 @@ trait DValueCodecs {
       def fromString(s: String): Option[Path] =
         Some(Path.fromString(s))
 
-      def typeDefinition:StringDefinition =
+      def typeDefinition: StringDefinition =
         StringDefinition.empty
     }
